refactor: modularize init shop setup
