+++ scripts/src/runtime.ts
@@ -1,0 +1,33 @@
+import { execSync } from "node:child_process";
+
+/**
+ * Ensure that the runtime meets the minimum supported versions for Node.js and pnpm.
+ * Exits the process with an error if requirements are not met.
+ */
+export function ensureRuntime(): void {
+  const nodeMajor = Number(process.version.replace(/^v/, "").split(".")[0]);
+  if (nodeMajor < 20) {
+    console.error(
+      `Node.js v20 or later is required. Current version: ${process.version}`
+    );
+    process.exit(1);
+  }
+
+  let pnpmVersion: string;
+  try {
+    pnpmVersion = execSync("pnpm --version", { encoding: "utf8" }).trim();
+  } catch {
+    console.error(
+      "Failed to determine pnpm version. pnpm v10 or later is required."
+    );
+    process.exit(1);
+  }
+
+  const pnpmMajor = Number(pnpmVersion.split(".")[0]);
+  if (pnpmMajor < 10) {
+    console.error(
+      `pnpm v10 or later is required. Current version: ${pnpmVersion}`
+    );
+    process.exit(1);
+  }
+}
